style: move admin user details to their own class
