+++ src/main/java/edu/kit/hci/soli/config/security/SoliUserDetails.java
@@ -30,31 +30,4 @@ public interface SoliUserDetails extends UserDetails {
     }
 
     User getUser();
-
-    class Admin implements SoliUserDetails {
-        private final User user;
-        private final String password;
-
-        public Admin(User user, String password) {
-            this.user = user;
-            this.password = password;
-        }
-
-        @Override
-        public List<SimpleGrantedAuthority> getAuthorities() {
-            List<SimpleGrantedAuthority> result = SoliUserDetails.super.getAuthorities();
-            result.add(new SimpleGrantedAuthority("ROLE_ADMIN"));
-            return result;
-        }
-
-        @Override
-        public String getPassword() {
-            return password;
-        }
-
-        @Override
-        public User getUser() {
-            return user;
-        }
-    }
 }
